+++ services/telegram.js
@@ -1,21 +1,33 @@
 process.env.NTBA_FIX_319 = 1;
 const TelegramBot = require('node-telegram-bot-api');
 
+const serviceName = 'telegram';
+var internal, client;
+
 async function onMessage(msg) {
     const chatId = msg.chat.id;
     const isDM = msg.chat.type == 'private';
     const lang = msg.from.language_code;
-    const bot = await internal.simabot(msg);
+    const crossmsg = {
+        serviceName: serviceName,
+        userid: msg.from.id,
+        nickname: '@' + msg.from.username,
+        channelId: msg.chat.id,
+        isDM: isDM,
+        content: msg.text,
+        lang: lang,
+        msg: msg
+    };
+    var out = await internal.simabot(crossmsg);
     const messageId = msg.message_id;
-    if (bot) {
-        var out = bot.substring(0, 4095);
-        if (out.startsWith('!!!')) {
-            try {
-
-            } catch (err) {
+    if (out) {
+        if (typeof out == 'object') {
+            if (out.action == 'remove') {
                 client.deleteMessage(chatId, messageId);
             }
-        } else {
+        }
+        if (typeof out == 'string') {
+            out = out.substring(0, 4095);
             client.sendMessage(chatId, out);
         }
     }
@@ -38,6 +50,7 @@ exports.sendNotify = function (msg, isDM) {
 }
 
 exports.init = function (token, data) {
-    const client = new TelegramBot(token, { polling: true });
+    internal = data;
+    client = new TelegramBot(token, { polling: true });
     client.on('message', onMessage);
 }